Refactor computemanagementservice
Move methods into servicemanagementservice and delete the
computemanagementservice class. In addition rename the
replication and share methods to follow the naming schema
